Switch youtube API endpoint to text response
